lab9/components/mlp_trainer/requirements.txt
google-cloud-storage
pandas
keras
tensorflow
h5py
scikit-learn